region access point for index localftp and major internal refactoring
I basically split the data from the index class fetchers into their own files. This was necessary to more precisely define the available access points and dataset for all the fetchers.
