+++ argopy/data_fetchers/localftp_data.py
@@ -43,7 +43,7 @@ from .proto import ArgoDataFetcherProto
 from argopy.errors import NetCDF4FileNotFoundError
 from argopy.utilities import list_standard_variables, load_dict, mapp_dict
 from argopy.options import OPTIONS
-from argopy.stores import filestore, indexstore, indexfilter_wmo
+from argopy.stores import filestore, indexstore, indexfilter_wmo, indexfilter_box
 
 access_points = ['wmo']
 exit_formats = ['xarray']
@@ -209,10 +209,6 @@ class LocalFTPArgoDataFetcher(ArgoDataFetcherProto):
     def cachepath(self):
         """ Return path to cache file for this request """
         return [self.fs.cachepath(file) for file in self.files]
-
-    def clear_cache(self):
-        """ Remove cache files and entries from resources open with this fetcher """
-        return self.fs.clear_cache()
 
     def xload_multiprof(self, ncfile: str):
         """Load an Argo multi-profile file as a collection of points
@@ -423,139 +419,3 @@ class Fetch_wmo(LocalFTPArgoDataFetcher):
                     for cyc in self.CYC:
                         self._list_of_argo_files.append(self._absfilepath(wmo, cyc, errors=errors))
         return self
-
-
-class LocalFTPArgoIndexFetcher(ABC):
-    """ Manage access to Argo index from a local copy of GDAC ftp
-
-    """
-    ###
-    # Methods to be customised for a specific request
-    ###
-    @abstractmethod
-    def init(self):
-        """ Initialisation for a specific fetcher """
-        pass
-
-    @abstractmethod
-    def cname(self):
-        """ Return a unique string defining the request """
-        pass
-
-    @abstractmethod
-    def cachepath(self):
-        pass
-
-    @abstractmethod
-    def filter_index(self):
-        """ Custom search in the argo index file
-
-        Returns
-        -------
-        Instance of index_filter_proto
-        """
-        pass
-
-    ###
-    # Methods that must not change
-    ###
-    def __init__(self,
-                 local_ftp: str = "",
-                 index_file: str = "ar_index_global_prof.txt",
-                 cache: bool = False,
-                 cachedir: str = "",
-                 **kwargs):
-        """ Init fetcher
-
-            Parameters
-            ----------
-            local_path : str
-                Path to the directory with the 'dac' folder and index file
-        """
-        self.cache = cache
-        self.definition = 'Local ftp Argo index fetcher'
-        self.local_ftp = OPTIONS['local_ftp'] if local_ftp == '' else local_ftp
-        self.index_file = index_file
-        self.fs = indexstore(cache, cachedir, os.path.sep.join([self.local_ftp, self.index_file]))
-        self.init(**kwargs)
-
-    def __repr__(self):
-        summary = ["<indexfetcher '%s'>" % self.definition]
-        summary.append("FTP: %s" % self.local_ftp)
-        summary.append("Domain: %s" % self.cname())
-        return '\n'.join(summary)
-
-    def to_dataframe(self):
-        """ filter local index file and return a pandas dataframe """
-        df = self.fs.open_dataframe(self.filter_index())
-
-        # Post-processing of the filtered index:
-        df['wmo'] = df['file'].apply(lambda x: int(x.split('/')[1]))
-
-        # institution & profiler mapping for all users
-        # todo: may be we need to separate this for standard and expert users
-        institution_dictionnary = load_dict('institutions')
-        df['tmp1'] = df.institution.apply(lambda x: mapp_dict(institution_dictionnary, x))
-        df = df.rename(columns={"institution": "institution_code", "tmp1": "institution"})
-
-        profiler_dictionnary = load_dict('profilers')
-        df['profiler'] = df.profiler_type.apply(lambda x: mapp_dict(profiler_dictionnary, int(x)))
-        df = df.rename(columns={"profiler_type": "profiler_code"})
-
-        return df
-
-    def to_xarray(self):
-        """ Load Argo index and return a xarray Dataset """
-        return self.to_dataframe().to_xarray()
-
-    def clear_cache(self):
-        """ Remove cache files and entries from resources open with this fetcher """
-        return self.fs.clear_cache()
-
-
-class IndexFetcher_wmo(LocalFTPArgoIndexFetcher):
-    """ Manage access to local ftp Argo data for: a list of WMOs
-
-    """
-    def init(self, WMO: list = [], CYC=None, **kwargs):
-        """ Create Argo data loader for WMOs
-
-            Parameters
-            ----------
-            WMO : list(int)
-                The list of WMOs to load all Argo data for.
-            CYC : int, np.array(int), list(int)
-                The cycle numbers to load.
-        """
-        if isinstance(WMO, int):
-            WMO = [WMO]  # Make sure we deal with a list
-        if isinstance(CYC, int):
-            CYC = np.array((CYC,), dtype='int')  # Make sure we deal with an array of integers
-        if isinstance(CYC, list):
-            CYC = np.array(CYC, dtype='int')  # Make sure we deal with an array of integers
-        self.WMO = WMO
-        self.CYC = CYC
-        self.fcls = indexfilter_wmo(self.WMO, self.CYC)
-
-    @property
-    def cachepath(self):
-        return self.fs.cachepath(self.fcls.uri())
-
-    def cname(self):
-        """ Return a unique string defining the request
-
-        """
-        return self.fcls.uri()
-
-    def filter_index(self):
-        """ Search for one or more WMO in the argo index file
-
-        Parameters
-        ----------
-        index_file: _io.TextIOWrapper
-
-        Returns
-        -------
-        csv rows matching the request, as a in-memory string. Or None.
-        """
-        return self.fcls
