+++ argopy/stores/__init__.py
@@ -1,4 +1,4 @@
-from .argo_index import indexstore, indexfilter_wmo
+from .argo_index import indexstore, indexfilter_wmo, indexfilter_box
 from .fsspec_wrappers import filestore, httpstore, memorystore
 
 #
@@ -6,6 +6,7 @@ __all__ = (
     # Classes:
     "indexstore",
     "indexfilter_wmo",
+    "indexfilter_box",
     "filestore",
     "httpstore",
     "memorystore"
